+++ sim/agent/testagent/AntFindFoodAntTest.hpp
@@ -17,7 +17,7 @@ namespace cdc
 		AntFindFoodAntTest(GuiEventManager& manager, AntHome& home, NavGraphHelper& navGraphHelper, const Node& startNode);
 		virtual ~AntFindFoodAntTest();
 
-		virtual void update(long ticks, const Percept& percept) override;
+		virtual void update(uint ticks, const Percept& percept) override;
 
 		virtual bool isGoalFinished() const override;
