+++ sim/agent/testagent/AntGoalTester.hpp
@@ -13,7 +13,7 @@ namespace cdc
 		AntGoalTester(GuiEventManager& manager, AntHome& home, NavGraphHelper& navGraphHelper, const Node& node) :
 			Ant(manager, home, navGraphHelper, node) {}
 		virtual ~AntGoalTester() {}
-		virtual void update(long ticks, const Percept& percept) override {}
+		virtual void update(uint ticks, const Percept& percept) override {}
 		virtual bool isGoalFinished() const = 0;
 	};
 }
